Moved /script folder to /framework/script
