+++ src/framework/script/script-registry.js
@@ -1,9 +1,9 @@
-import { EventHandler } from '../core/event-handler.js';
+import { EventHandler } from '../../core/event-handler.js';
 
 /* eslint-disable-next-line no-unused-vars */
 import { ScriptType } from './script-type.js';
 
-/** @typedef {import('../framework/app-base.js').AppBase} AppBase */
+/** @typedef {import('../app-base.js').AppBase} AppBase */
 
 /**
  * Container for all {@link ScriptType}s that are available to this application. Note that
